LIVE-2187 polish portfolio + graphs
